revert: 重构加密流程
